refactor: extracted generateType, generateBuiltinType
