+++ packages/gatsby-transformer-raml/src/utils/type/generate-built-in-type.js
@@ -1,0 +1,13 @@
+function generateBuiltinType(property) {
+  switch (property.builtinType) {
+    case 'date-only':
+    case 'time-only':
+    case 'datetime-only':
+    case 'datetime':
+      return 'string';
+    default:
+      return property.builtinType;
+  }
+}
+
+module.exports = generateBuiltinType;
